// ui.frontend/src/types/ContentType.ts
export type ContentType = 'page' | 'xf' | 'asset' | 'cf' | 'child' | 'custom' | 'all';
export type ContentTypeLabel =
  | 'Page'
  | 'XF (Experience Fragment)'
  | 'DAM Asset'
  | 'CF (Content Fragment)'
  | 'Child (Component)'
  | 'Custom Node Type'
  | 'All Node Types';

export type ContentTypeProperty = 'cq:Page' | 'dam:Asset' | 'nt:unstructured' | 'custom' | 'nt:base';